requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.12
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-gis==0.18
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
